update test file name
