+++ rust/tests/read_delta_log_test.rs
@@ -1,0 +1,96 @@
+use deltalake::DeltaTableBuilder;
+use pretty_assertions::assert_eq;
+use std::collections::HashMap;
+use std::time::SystemTime;
+
+#[allow(dead_code)]
+mod fs_common;
+
+#[tokio::test]
+async fn test_log_buffering() {
+    let n_commits = 10;
+    let path = "./tests/data/simple_table_with_no_checkpoint";
+    let mut table = fs_common::create_table(path, None).await;
+    for _ in 0..n_commits {
+        let a = fs_common::add(3 * 60 * 1000);
+        fs_common::commit_add(&mut table, &a).await;
+    }
+
+    let max_iter = 10;
+    let buf_size = 10;
+
+    let location = deltalake::builder::ensure_table_uri(path).unwrap();
+    
+    // use storage that sleeps 10ms on every `get`
+    let store = std::sync::Arc::new(
+        fs_common::SlowStore::new(
+            location.clone(),
+            deltalake::storage::config::StorageOptions::from(HashMap::new()
+        )
+    ).unwrap());
+
+    let mut seq_version = 0;
+    let t = SystemTime::now();
+    for _x in 0..max_iter {
+        let mut table_seq = DeltaTableBuilder::from_uri(path).with_storage_backend(store.clone(), location.clone()).with_version(0).with_buffer(1).unwrap().load().await.unwrap();
+        table_seq.update_incremental(None).await.unwrap();
+        seq_version = table_seq.version();
+    }
+    let time_seq = t.elapsed().unwrap();
+
+    let mut buf_version = 0;
+    let t2 = SystemTime::now();
+    for _x in 0..max_iter {
+        let mut table_buf = DeltaTableBuilder::from_uri(path).with_storage_backend(store.clone(), location.clone()).with_version(0).with_buffer(buf_size).unwrap().load().await.unwrap();
+        table_buf.update_incremental(None).await.unwrap();
+        buf_version = table_buf.version();
+    }
+    let time_buf = t2.elapsed().unwrap();
+
+    // buffered was at least 2x faster
+    assert!(time_buf.as_secs_f64() < 0.5*time_seq.as_secs_f64());
+
+    // updates match
+    assert_eq!(seq_version, 10);
+    assert_eq!(buf_version, 10);
+}
+
+#[tokio::test]
+async fn test_log_buffering_success_explicit_version() {
+    let n_commits = 10;
+    let path = "./tests/data/simple_table_with_no_checkpoint_2";
+    let mut table = fs_common::create_table(path, None).await;
+    for _ in 0..n_commits {
+        let a = fs_common::add(3 * 60 * 1000);
+        fs_common::commit_add(&mut table, &a).await;
+    }
+    let buf_sizes = [1,2,10,50];
+    for buf_size in buf_sizes {
+        let mut table = DeltaTableBuilder::from_uri(path).with_version(0).with_buffer(buf_size).unwrap().load().await.unwrap();
+        table.update_incremental(None).await.unwrap();
+        assert_eq!(table.version(), 10);
+        
+        let mut table = DeltaTableBuilder::from_uri(path).with_version(0).with_buffer(buf_size).unwrap().load().await.unwrap();
+        table.update_incremental(Some(0)).await.unwrap();
+        assert_eq!(table.version(), 10);
+
+        let mut table = DeltaTableBuilder::from_uri(path).with_version(0).with_buffer(buf_size).unwrap().load().await.unwrap();
+        table.update_incremental(Some(1)).await.unwrap();
+        assert_eq!(table.version(), 1);
+
+        let mut table = DeltaTableBuilder::from_uri(path).with_version(0).with_buffer(buf_size).unwrap().load().await.unwrap();
+        table.update_incremental(Some(10)).await.unwrap();
+        assert_eq!(table.version(), 10);
+
+        let mut table = DeltaTableBuilder::from_uri(path).with_version(0).with_buffer(buf_size).unwrap().load().await.unwrap();
+        table.update_incremental(Some(20)).await.unwrap();
+        assert_eq!(table.version(), 10);
+    }
+}
+
+#[tokio::test]
+async fn test_log_buffering_fail() {
+    let path = "./tests/data/simple_table_with_no_checkpoint";
+    let table_err = DeltaTableBuilder::from_uri(path).with_version(0).with_buffer(0).is_err();
+    assert!(table_err);
+}
